+++ projects/ids-enterprise-ng/src/lib/toolbar-flex/soho-toolbar-flex.spec.ts
@@ -2,14 +2,17 @@
 
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { Component, DebugElement, ViewChild } from '@angular/core';
+import {Component, DebugElement, Input, ViewChild} from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import {SohoToolbarFlexComponent} from './soho-toolbar-flex.component';
+import {SohoToolbarFlexComponent, SohoToolbarFlexSearchFieldComponent} from './soho-toolbar-flex.component';
 import {SohoToolbarFlexModule} from './soho-toolbar-flex.module';
 import {SohoMenuButtonModule} from '../menu-button';
+import {TestHelper} from '../utils';
 
 @Component({
   template: `<soho-toolbar-flex
+    title="Toolbar With Ajax 'More Actions' Menu"
+    [moreMenuBeforeOpenFunction]="onBeforeMoreMenuOpen"
     (selected)="onSelected($event)">
 
     <soho-toolbar-flex-section [isTitle]="true">
@@ -45,6 +48,7 @@ import {SohoMenuButtonModule} from '../menu-button';
 
     <soho-toolbar-flex-section [isSearch]="true">
       <input soho-toolbar-flex-searchfield
+             id="search-field"
              placeholder="Search..."
              [clearable]="true"
              [collapsible]="false"
@@ -62,6 +66,7 @@ import {SohoMenuButtonModule} from '../menu-button';
 })
 class SohoToolbarFlexTestComponent {
   @ViewChild(SohoToolbarFlexComponent) toolbarFlex: SohoToolbarFlexComponent;
+  @ViewChild(SohoToolbarFlexSearchFieldComponent) searchField: SohoToolbarFlexSearchFieldComponent;
 
   private MENU_RESPONSE_HTML = `
     <li><a href="#" id="AJAXOptionOne" data-action="{'button':'AJAX Option 1'}">AJAX Option 1</a></li>
@@ -120,9 +125,13 @@ class SohoToolbarFlexTestComponent {
     }
     console.log(data);
   }
+
+  onChange(event: SohoSearchFieldEvent) {
+    console.log('Search Changed ' + this.model.searchValue);
+  }
 }
 
-describe('Soho Toolbar Flex Render', () => {
+describe('Soho Toolbar Flex Tests', () => {
   let toolbarflex: SohoToolbarFlexComponent;
   let component: SohoToolbarFlexTestComponent;
   let fixture: ComponentFixture<SohoToolbarFlexTestComponent>;
@@ -209,6 +218,63 @@ describe('Soho Toolbar Flex Render', () => {
       subItem.click();
       expect(spy.calls.count()).toEqual(1);
     });
+
+    it('search field', () => {
+      fixture.detectChanges();
+
+      const searchField = de.query(By.css('#search-field')).nativeElement;
+
+      TestHelper.testFireEvent(searchField, 'selected', component.searchField['selected']);
+      TestHelper.testFireEvent(searchField, 'cleared', component.searchField['cleared']);
+    });
   });
 
+  describe('Check Inputs', () => {
+    it('flex toolbar ', () => {
+      fixture.detectChanges();
+
+      expect(component.toolbarFlex.options.beforeMoreMenuOpen).toBeDefined('more menu callback not defined');
+
+      const spy = spyOn<any>(component.toolbarFlex, 'updated').and.callThrough();
+      component.toolbarFlex.moreMenuBeforeOpenFunction = undefined;
+      fixture.detectChanges();
+
+      expect(component.toolbarFlex.options.beforeMoreMenuOpen).toBeUndefined('more menu callback is still defined');
+      expect(spy.calls.count()).toEqual(1);
+    });
+
+    it('flex toolbar searchfield', () => {
+      fixture.detectChanges();
+
+      expect(component.searchField).toBeDefined('search field is undefined');
+      expect(component.searchField.options.clearable).toBeTruthy('search field not clearable');
+      expect(component.searchField.options.collapsible).toBeFalsy('search field collapsible');
+      expect(component.searchField.options.filterMode).toBe('contains');
+      expect(component.searchField.options.collapsibleOnMobile).toBeUndefined();
+      expect(component.searchField.options.source).toBeUndefined();
+      expect(component.searchField.options.template).toBeUndefined();
+
+      const spy = spyOn<any>(component.searchField.searchField, 'updated').and.callThrough();
+
+      component.searchField.options.filterMode = 'startsWith';
+      component.searchField.source = (query, done) => {
+        this.objectBasedData().subscribe((items) => {
+          done(query, items);
+        });
+      };
+      component.searchField.clearable = false;
+      component.searchField.collapsible = true;
+      component.searchField.collapsibleOnMobile = true;
+      fixture.detectChanges();
+      expect(spy.calls.count()).toEqual(1);
+
+      expect(component.searchField.options.clearable).toBeFalsy('search field clearable');
+      expect(component.searchField.options.collapsible).toBeTruthy('search field not collapsible');
+      expect(component.searchField.options.collapsibleOnMobile).toBeTruthy('search field not collapsible on mobile');
+      expect(component.searchField.options.filterMode).toBe('startsWith');
+      expect(component.searchField.options.source).toBeDefined();
+
+    });
+
+  });
 });